Add basic admin check on accounts to be disabled.
Adds a check that is enabled by default that will make sure any account to be disabled does not have adminCount set > 0. This is a very basic admin check, better than nothing. Failing this check will increment the error counter and log the error appropriately. It can be disabled by setting a bool in the main file like all other options.